Fixed a crash when start menu was open
The crash was due to error-checking on a SetForegroundWindow that would fail when trying to open the TrayFolder pop-up while start menu (or other types of pop-ups) were already open.

I've decided to let the call to fail, since I haven't found a way to make it work (don't want to mess with Windows input messages, etc.) Even though it means that when that happens, the user can no longer dismiss the pop-up by clicking outside of it. To ease that a bit, I've added an explicit "Dismiss" entry to the pop-up.

Interestingly, having a debugger attached changes this behavior (SetForegroundWindow doesn't fail when running under debugger!)

Also, added sorting of the list (dirs first, and case-insensitive by title).

Another addition is writing the logs to debugger output (already minimal).